Set the min Go version to 1.7 in go.mod
`go` tool requires the minimal Go version to be specified in the
go.mod file and adds it if it's not there yet. As a result it sets git
workspace in dirty state and does not allow to execute `go` tools with
`-mod=readonly`.

go-sensor supports Go version 1.7 and above. Although modules were
introduced in 1.11 it still makes sense to have the officially
supported Go version listed in go.mod.