Don't fail CI builds if 1.7 does not pass
`github.com/looplab/fsm@master` has recently dropped support of 1.7 by
introducing the use of `sort.Slice` that was only added in 1.8. As a
result any app that uses go1.7 and does not have its dependencies
vendored will raise a compile-time error. A suggested temporary
workaround would be to vendor `github.com/looplab/fsm` and lock it to
v0.1.0