Fix data race while stopping the autotrace/internal.Timer
Both `(internal.Timer).intervalDone` and
`(internal.Timer).intervalTicker` are initialized in a goroutine after
the delay has passed. If `(*internal.Timer).Stop()` is called shortly
after the delay has passed, but before the ticker is initialized,
there is a chance that the `t.intervalTicker != nil` check will be
evaluated to `false`, but the ticker will still be created. As a
result, the goroutine that executes `job()` will leak without any way
to stop it.