(instana.Tracer).Extract() return `nil` trace context in case of an error
The OpenTracing documentation encourages to ignore the error returned
by `(opentracing.Tracer).Extract()` and rely on the `SpanContext`
being `nil` in this case. This contradicts the error-handling
convention in Go that does not guaranteed the returned values to be
useful if an error is returned. To address this,
`(instana.Tracer).Extract()` now explicitly returns `nil` along with
an error.