do not discard the Instana tracestate list member
Previously, when more than 32 list members were in the incoming
tracestate header value, and the Instana-specific list member (in=...)
was at position 33 or later, we would drop it together with the other
list members that need to be droppped. With this change, we give the
Instana list member higher priority than list members from other
vendors.

We also keep it separate in the TraceState struct.

Furthermore, this simplifies/reduces the API of the TraceState module.
We do not have a use case for adding/removing list members for arbitrary
vendors - we only need to propagate the list members we received and
make sure that the Instana trace state value is added in the left-most
position.