feat: improve robustness of custom service naming
Use INSTANA_SERVICE_NAME value for span.data.service annotation.

Previously, span.data.service has only been added when configured via in-code
configuration, but not when the environment variable INSTANA_SERVICE_NAME is
set. The Go tracer implicitly relied on the back end to link spans to
infrastructure entities and get the service name from there. Adding the
annotation to the span makes service name extraction more resilient, in
particular with respect to situations in which spans cannot be linked to
infrastructure.

The binary name is now kept as a separate attribute in the sensor struct. This
is to keep the pre-existing behavior to fall back to the binary's name when
sending the infrastructure snapshot `name` attribute and the `service`
attribute for events. However, we explicitly do not want to add the
binary's name as an annotation to all spans.